fix: move CORS handling by introducing setupCors middleware
